+++ src/components/Export/index.js
@@ -1,0 +1,5 @@
+/**
+ * Re-export to prevent nested path
+ */
+import Export from './Export'
+export default Export
